requirements.txt
fastapi
uvicorn[standard]
openai==0.27.10
faiss-cpu
python-dotenv
tiktoken
langchain
pydantic
httpx
redis
pytest